moved server stuff to its own folder
